Update itsystem migration for `atomic = False`
We get an error from CircleCI running the migrations because
of version of SQLITE3 being used. The error is because we
rename a database table, `itsystems_systeminstance` to `itsystems_system`
and the database table is in a foreign key. We follow recommendation from
error message:

"django.db.utils.NotSupportedError: Renaming the 'itsystems_systeminstance' table while in a transaction is not supported on SQLite < 3.26 because it would break referential integrity. Try adding `atomic = False` to the Migration class."

A better solution might be refactor the migrations to name the table name
earlier in the migrations.